UPSTREAM: <carry>: Short-circuit HPA oapi/v1.DC
The legacy oapi v1 group-version very much confuses anything not
designed to explicitly work with it.  Since we now don't do any custom
HPA setup, we need to teach the scale client and the HPA what to do
with the oapi version of DC, since it won't even show up in its
discovery process.

Origin-commit: eea8aa3fcc1b9164da0646be9b3461444da1d3af

Kubernetes-commit: 8ebbd8260651992a9f7f1ffea197ae7246586179