UPSTREAM: <carry>: Short-circuit HPA oapi/v1.DC
The legacy oapi v1 group-version very much confuses anything not
designed to explicitly work with it.  Since we now don't do any custom
HPA setup, we need to teach the scale client and the HPA what to do
with the oapi version of DC, since it won't even show up in its
discovery process.

Kubernetes-commit: 00f2f88d0f3faa5cde4e8a5abea5b73777ded75a